third_party/WebKit/LayoutTests/flag-specific/enable-blink-features=LayoutNG/external/wpt/css/css-shapes/spec-examples/shape-outside-017-expected.txt

This is a testharness.js-based test.
FAIL CSS Test: Shape from box value - content-box Uncaught Error: assert_approx_equals: Line 0 is positioned properly expected 193 +/- 2 but got 185
Harness: the test ran to completion.
